Changes:
• Refactored code modules Graphics, Entity Component, Resource from SchoolGameEngine project to work with Gem.
• Classes are not ready to be used yet.

+++ Project/Common/Core/ValueTypes.h
@@ -17,11 +17,11 @@ namespace Gem
 
     typedef unsigned char       UInt8;
     typedef unsigned short      UInt16;
-    typedef unsigned long       UInt32;
+    typedef unsigned int        UInt32;
     typedef unsigned long long  UInt64;
     typedef signed char         SInt8;
     typedef signed short        SInt16;
-    typedef signed long         SInt32;
+    typedef signed int          SInt32;
     typedef signed long long    SInt64;
     typedef float               Float32;
     typedef double              Float64;
